ai-dashboard/ai-dashboard/ai-analytics/backend/requirements.txt
# ------------------------
# Core Backend
# ------------------------
fastapi
uvicorn

# ------------------------
# Data handling
# ------------------------
pandas
openpyxl

# ------------------------
# LangGraph + LangChain
# ------------------------
langgraph
langchain
langchain-google-genai

# ------------------------
# Environment management
# ------------------------
python-dotenv

# ------------------------
# Optional (recommended)
# ------------------------
typing-extensions
